+++ Audibly.App/App.xaml.cs
@@ -1,5 +1,5 @@
 ﻿// Author: rstewa · https://github.com/rstewa
-// Updated: 02/26/2025
+// Updated: 03/02/2025
 
 using System;
 using System.Collections.Generic;