+++ Audibly.App/AppShell.xaml.cs
@@ -1,5 +1,5 @@
 ﻿// Author: rstewa · https://github.com/rstewa
-// Updated: 02/20/2025
+// Updated: 03/02/2025
 
 using System;
 using System.Collections.Generic;
@@ -12,11 +12,9 @@ using Audibly.App.Services;
 using Audibly.App.ViewModels;
 using Audibly.App.Views;
 using CommunityToolkit.WinUI;
-using Microsoft.UI.Composition.SystemBackdrops;
 using Microsoft.UI.Xaml;
 using Microsoft.UI.Xaml.Controls;
 using Microsoft.UI.Xaml.Input;
-using Microsoft.UI.Xaml.Media;
 using Microsoft.UI.Xaml.Navigation;
 using Constants = Audibly.App.Helpers.Constants;
 using DispatcherQueue = Microsoft.UI.Dispatching.DispatcherQueue;
@@ -45,12 +43,8 @@ public sealed partial class AppShell : Page
 
         // set the title bar
         var window = WindowHelper.GetMainWindow();
-        if (window != null)
-        {
-            window.SetTitleBar(AppTitleBar);
-            //window.SizeChanged += Window_SizeChanged; // Subscribe to the SizeChanged event
-        }
-
+        if (window != null) window.SetTitleBar(AppTitleBar);
+        //window.SizeChanged += Window_SizeChanged; // Subscribe to the SizeChanged event
         AppShellFrame.Navigate(typeof(LibraryCardPage));
 
         Loaded += (_, _) => { NavView.SelectedItem = LibraryCardMenuItem; };